added login, logout  and register

+++ study/myapp/views.py
@@ -1,9 +1,60 @@
 from django.shortcuts import render, redirect
+from django.contrib import messages
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room, Topic
 from .forms import RoomForm
 
+
 # Create your views here.
+def login_page(request):
+    page = "login"
+    if request.user.is_authenticated:
+        return redirect("home")
+
+    if request.method == "POST":
+        username = request.POST.get("username").lower()
+        password = request.POST.get("password")
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, "User doesn't exist")
+        
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "Username or password does not exist")
+
+    context = {"page":page}
+    return render(request, "myapp/register_login.html", context)
+
+def logout_user(request):
+    logout(request)
+    return redirect('home')
+
+def register_page(request):
+    form = UserCreationForm()
+
+    if request.method=="POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "An error occurred during registration")
+
+    return render(request, "myapp/register_login.html", {"form": form})
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') else ''
@@ -16,11 +67,14 @@ def home(request):
     context = {"rooms": rooms, "topics": topics, "room_count": room_count}
     return render(request, "myapp/home.html", context)
 
+
+
 def room(request, pk):
     room = Room.objects.get(id=pk) 
     context = {"room":  room}
     return render(request, 'myapp/room.html', context)
 
+@login_required(login_url="login")
 def create_room(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -31,9 +85,13 @@ def create_room(request):
     context = {"form": form}
     return render(request, 'myapp/room_form.html', context)
 
+@login_required(login_url="login")
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed to perform this action.")
     
     if request.method=="POST":
         form = RoomForm(request.POST, instance=room)
@@ -44,8 +102,13 @@ def update_room(request, pk):
     context = {"form" : form}
     return render(request, 'myapp/room_form.html', context)
 
+@login_required(login_url="login")
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed to perform this action.")
+    
     if request.method == "POST":
         room.delete()
         return redirect("home")
